+++ framework/db/ActiveRecord.php
@@ -445,15 +445,14 @@ class ActiveRecord extends BaseActiveRecord
                 if ($this->getAttribute($name) === null) {
                     $id = $db->getLastInsertID($table->sequenceName);
                     $this->setAttribute($name, $id);
-                    $this->setOldAttribute($name, $id);
+                    $values[$name] = $id;
                     break;
                 }
             }
         }
-        foreach ($values as $name => $value) {
-            $this->setOldAttribute($name, $value);
-        }
+
         $this->afterSave(true);
+        $this->setOldAttributes($values);
 
         return true;
     }