+++ framework/log/Dispatcher.php
@@ -50,6 +50,12 @@ use yii\base\Component;
  * Yii::$app->log->targets['file']->enabled = false;
  * ~~~
  *
+ * @property integer $flushInterval How many messages should be logged before they are sent to targets. This
+ * method returns the value of [[Logger::flushInterval]].
+ * @property Logger $logger The logger. If not set, [[\Yii::getLogger()]] will be used.
+ * @property integer $traceLevel How many application call stacks should be logged together with each message.
+ * This method returns the value of [[Logger::traceLevel]]. Defaults to 0.
+ *
  * @author Qiang Xue <qiang.xue@gmail.com>
  * @since 2.0
  */
@@ -61,9 +67,25 @@ class Dispatcher extends Component
      */
     public $targets = [];
     /**
-     * @var Logger the logger. If not set, [[\Yii::getLogger()]] will be used.
+     * @var Logger the logger.
      */
-    public $logger;
+    private $_logger;
+
+
+    /**
+     * @inheritdoc
+     */
+    public function __construct($config = [])
+    {
+        if (isset($config['logger'])) {
+            $this->setLogger($config['logger']);
+            unset($config['logger']);
+        }
+        // connect logger and dispatcher
+        $this->getLogger();
+
+        parent::__construct($config);
+    }
 
     /**
      * Initializes the logger by registering [[flush()]] as a shutdown function.
@@ -77,11 +99,30 @@ class Dispatcher extends Component
                 $this->targets[$name] = Yii::createObject($target);
             }
         }
+    }
 
-        if ($this->logger === null) {
-            $this->logger = Yii::getLogger();
+    /**
+     * Gets the connected logger.
+     * If not set, [[\Yii::getLogger()]] will be used.
+     * @property Logger the logger. If not set, [[\Yii::getLogger()]] will be used.
+     * @return Logger the logger.
+     */
+    public function getLogger()
+    {
+        if ($this->_logger === null) {
+            $this->setLogger(Yii::getLogger());
         }
-        $this->logger->dispatcher = $this;
+        return $this->_logger;
+    }
+
+    /**
+     * Sets the connected logger.
+     * @param Logger $value the logger.
+     */
+    public function setLogger($value)
+    {
+        $this->_logger = $value;
+        $this->_logger->dispatcher = $this;
     }
 
     /**
@@ -90,7 +131,7 @@ class Dispatcher extends Component
      */
     public function getTraceLevel()
     {
-        return $this->logger->traceLevel;
+        return $this->getLogger()->traceLevel;
     }
 
     /**
@@ -101,7 +142,7 @@ class Dispatcher extends Component
      */
     public function setTraceLevel($value)
     {
-        $this->logger->traceLevel = $value;
+        $this->getLogger()->traceLevel = $value;
     }
 
     /**
@@ -110,7 +151,7 @@ class Dispatcher extends Component
      */
     public function getFlushInterval()
     {
-        return $this->logger->flushInterval;
+        return $this->getLogger()->flushInterval;
     }
 
     /**
@@ -123,7 +164,7 @@ class Dispatcher extends Component
      */
     public function setFlushInterval($value)
     {
-        $this->logger->flushInterval = $value;
+        $this->getLogger()->flushInterval = $value;
     }
 
     /**